125180 UInt32 for SetFIeldCounter

+++ BCT.Common.Messages.CPP/include/AbstractAggregate.h
@@ -9,7 +9,7 @@ namespace Bct
          class AbstractAggregate
          {
          public:
-            virtual int16_t FieldSetCounter() = 0;
+            virtual uint32_t FieldSetCounter() = 0;
          };
       }
    }
